+++ devices/linkam_t95/adapters/__init__.py
@@ -16,7 +16,3 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 # *********************************************************************
-
-from devices import SimulatedLinkamT95
-
-linkam = SimulatedLinkamT95()
